+++ cpp/euler020/euler020.cpp
@@ -32,7 +32,7 @@ int main( int argc, char* argv[] )
 
     // Compute sum
     int sum = 0;
-    for( int i=0; i<result.size(); i++ ){
+    for( int i=0; i<(int)result.size(); i++ ){
         sum += (result[i] - '0');
     }
